src/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
python-dotenv>=1.0.0
# Include all requirements from agent/requirements.txt
praw>=7.7.0
requests>=2.31.0
crewai>=0.30.12
crewai-tools>=0.30.12
reddit-mcp>=0.1.0
mcpadapt[crewai]
google-generativeai>=0.5.4
browser-use 